+++ out/extension.js
@@ -121,8 +121,9 @@ exports.deactivate = deactivate;
 // }
 function ensureTerminalExists() {
     if (vscode_1.window.terminals.length === 0) {
-        vscode_1.window.showErrorMessage('No active terminals');
-        return false;
+        vscode_1.window.showInformationMessage('No active terminals - new terminal create for you ;)');
+        const terminal = vscode_1.window.createTerminal('Lorc');
+        terminal.show();
     }
     return true;
 }