+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.0.2
 
-- Readme ok
+- if no terminal is active, we creates one
 
 ## 1.0.1